feat: adopt new landing page
